Move streaming model definition to the main library for future re-use
